Upgrade to JPA2
- add profession et listing arena export
- add expiration date for actuEntity
- reopen inscription
- add calendar

+++ app/asptt-toulouse-natation-app/src/com/asptttoulousenatation/core/shared/actu/GetAllActuAction.java
@@ -5,33 +5,33 @@ import net.customware.gwt.dispatch.shared.Action;
 
 public class GetAllActuAction implements Action<GetAllActuResult> {
 	
-	private long limitStart;
-	private long limitEnd;
+	private int limitStart;
+	private int limitEnd;
 
 	public GetAllActuAction() {
 		limitStart = 0;
-		limitEnd = Long.MAX_VALUE;
+		limitEnd = Integer.MAX_VALUE;
 	}
 
-	public GetAllActuAction(long pLimitStart, long pLimitEnd) {
+	public GetAllActuAction(int pLimitStart, int pLimitEnd) {
 		super();
 		limitStart = pLimitStart;
 		limitEnd = pLimitEnd;
 	}
 
-	public long getLimitStart() {
+	public int getLimitStart() {
 		return limitStart;
 	}
 
-	public void setLimitStart(long pLimitStart) {
+	public void setLimitStart(int pLimitStart) {
 		limitStart = pLimitStart;
 	}
 
-	public long getLimitEnd() {
+	public int getLimitEnd() {
 		return limitEnd;
 	}
 
-	public void setLimitEnd(long pLimitEnd) {
+	public void setLimitEnd(int pLimitEnd) {
 		limitEnd = pLimitEnd;
 	}
 }
